[HER-1012] Apply suggestions
